Create README.md
Executive with a proven track record for rapidly building and deploying products. Exceptional ability to inspire loyalty and commitment among team members while driving for results. Over twenty years in the software industry with 21+ products and systems delivered. Known as an innovative, motivating, performance-focused leader with demonstrated ability not only in technology but also in customer and professional services, customer relationship management, sales, and medical diagnostic tools as well as Electronic Medical Records for both In-Patient and Out-Patient.

22+ years experience architecting, designing and implementing software as a software architect, server software designer, project lead, project manager, vice president COO and CEO.